Move logic to Crypttab class
