+++ test/y2storage/crypttab_test.rb
@@ -1,5 +1,6 @@
 #!/usr/bin/env rspec
-# Copyright (c) [2018] SUSE LLC
+
+# Copyright (c) [2018-2019] SUSE LLC
 #
 # All Rights Reserved.
 #
@@ -30,6 +31,8 @@ describe Y2Storage::Crypttab do
 
   let(:path) { File.join(DATA_PATH, crypttab_name) }
 
+  let(:crypttab_name) { "crypttab" }
+
   let(:scenario) { "empty_hard_disk_50GiB" }
 
   describe "#initialize" do
@@ -53,6 +56,176 @@ describe Y2Storage::Crypttab do
 
       it "sets an empty list of entries" do
         expect(subject.entries).to be_empty
+      end
+    end
+  end
+
+  describe "#save_encryption_names" do
+    before do
+      allow(Storage).to receive(:read_simple_etc_crypttab).and_return(storage_entries)
+    end
+
+    let(:scenario) { "gpt_encryption" }
+
+    let(:devicegraph) { Y2Storage::StorageManager.instance.staging }
+
+    let(:device) { devicegraph.find_by_name(device_name) }
+
+    let(:crypttab_entries) do
+      [
+        crypttab_entry("luks1", device_name, "password", [])
+      ]
+    end
+
+    let(:device_name) { "/dev/sda1" }
+
+    let(:storage_entries) { crypttab_entries.map(&:to_storage_value) }
+
+    shared_examples "swap encryption" do |encryption_method|
+      it "encrypts the device with #{encryption_method} method" do
+        subject.save_encryption_names(devicegraph)
+
+        expect(device.encryption.method.to_sym).to eq(encryption_method)
+      end
+
+      it "uses the crypttab name for the new encrytion device" do
+        subject.save_encryption_names(devicegraph)
+
+        expect(device.encryption.basename).to eq("luks1")
+      end
+    end
+
+    shared_examples "swap encryption methods" do
+      context "and the crypttab entry matches the random swap encryption method" do
+        let(:encryption_method) { Y2Storage::EncryptionMethod.find(:random_swap) }
+
+        before do
+          allow(encryption_method).to receive(:available?).and_return(true)
+        end
+
+        include_examples "swap encryption", :random_swap
+      end
+
+      context "and the crypttab entry matches the protected swap encryption method" do
+        let(:encryption_method) { Y2Storage::EncryptionMethod.find(:protected_swap) }
+
+        before do
+          allow(encryption_method).to receive(:available?).and_return(true)
+        end
+
+        include_examples "swap encryption", :protected_swap
+      end
+
+      context "and the crypttab entry matches the secure swap encryption method" do
+        let(:encryption_method) { Y2Storage::EncryptionMethod.find(:secure_swap) }
+
+        before do
+          allow(encryption_method).to receive(:available?).and_return(true)
+        end
+
+        include_examples "swap encryption", :secure_swap
+      end
+    end
+
+    context "when the device indicated in a crypttab entry is currently encrypted" do
+      let(:device_name) { "/dev/sda4" }
+
+      before do
+        allow(Y2Storage::EncryptionMethod).to receive(:for_crypttab).and_return(encryption_method)
+      end
+
+      context "and the crypttab entry does not match an encryption method for swap" do
+        let(:encryption_method) { nil }
+
+        it "saves the crypttab name on the encryption device" do
+          encryption_before = device.encryption
+
+          expect(device.encryption.crypttab_name).to be_nil
+
+          subject.save_encryption_names(devicegraph)
+
+          expect(device.encryption.sid).to eq(encryption_before.sid)
+          expect(device.encryption.crypttab_name).to eq("luks1")
+        end
+      end
+
+      context "and the crypttab entry matches a not available encryption type" do
+        let(:encryption_method) { Y2Storage::EncryptionMethod.find(:secure_swap) }
+
+        before do
+          allow(encryption_method).to receive(:available?).and_return(false)
+        end
+
+        it "does not modify the device" do
+          expect(device.encryption.method.to_sym).to eq(:luks1)
+
+          subject.save_encryption_names(devicegraph)
+
+          expect(device.encryption.method.to_sym).to eq(:luks1)
+        end
+      end
+
+      include_examples "swap encryption methods"
+    end
+
+    context "when the device indicated in a crypttab entry is not currently encrypted" do
+      let(:device_name) { "/dev/sda1" }
+
+      before do
+        allow(Y2Storage::EncryptionMethod).to receive(:for_crypttab).and_return(encryption_method)
+      end
+
+      context "and the crypttab entry does not match an encryption method for swap" do
+        let(:encryption_method) { nil }
+
+        it "does not encrypt the device" do
+          expect(device.encrypted?).to eq(false)
+
+          subject.save_encryption_names(devicegraph)
+
+          expect(device.encrypted?).to eq(false)
+        end
+      end
+
+      context "and the crypttab entry matches a not available encryption type" do
+        let(:encryption_method) { Y2Storage::EncryptionMethod.find(:secure_swap) }
+
+        before do
+          allow(encryption_method).to receive(:available?).and_return(false)
+        end
+
+        it "does not encrypt the device" do
+          expect(device.encrypted?).to eq(false)
+
+          subject.save_encryption_names(devicegraph)
+
+          expect(device.encrypted?).to eq(false)
+        end
+      end
+
+      include_examples "swap encryption methods"
+    end
+
+    context "when the device indicated in a crypttab entry is not found" do
+      let(:device_name) { "/dev/sdb1" }
+
+      # Mock the system lookup performed as last resort to find a device
+      before { allow(Y2Storage::BlkDevice).to receive(:find_by_any_name) }
+
+      it "does not fail" do
+        expect { subject.save_encryption_names(devicegraph) }.to_not raise_error
+      end
+    end
+
+    context "when an encrypted device is not indicated in any crypttab entry" do
+      let(:device_name) { "/dev/sda4" }
+
+      it "does not modify the crypttab name of that encryption device" do
+        device = devicegraph.find_by_name("/dev/sda5")
+
+        subject.save_encryption_names(devicegraph)
+
+        expect(device.encryption.crypttab_name).to be_nil
       end
     end
   end
