+++ test/y2storage/encryption_test.rb
@@ -27,161 +27,9 @@ describe Y2Storage::Encryption do
     fake_scenario(scenario)
   end
 
-  let(:scenario) { "gpt_encryption" }
-
   let(:devicegraph) { Y2Storage::StorageManager.instance.staging }
 
-  describe ".save_crypttab_names" do
-    before do
-      allow(Storage).to receive(:read_simple_etc_crypttab).and_return(storage_entries)
-    end
-
-    let(:device) { devicegraph.find_by_name(device_name) }
-
-    let(:crypttab_entries) do
-      [
-        crypttab_entry("luks1", device_name, "password", [])
-      ]
-    end
-
-    let(:device_name) { "/dev/sda1" }
-
-    let(:storage_entries) { crypttab_entries.map(&:to_storage_value) }
-
-    context "when a path to a crypttab file is given" do
-      it "tries to read the crypttab file" do
-        expect(Y2Storage::Crypttab).to receive(:new).and_call_original
-
-        described_class.save_crypttab_names(devicegraph, "/etc/crypttab")
-      end
-    end
-
-    context "when a crypttab object is given" do
-      it "uses the crypttab object and does not try to read a new one" do
-        crypttab = Y2Storage::Crypttab.new
-
-        expect(Y2Storage::Crypttab).to_not receive(:new)
-
-        described_class.save_crypttab_names(devicegraph, crypttab)
-      end
-    end
-
-    shared_examples "swap encryption" do |encryption_method|
-      it "encrypts the device with #{encryption_method} method" do
-        described_class.save_crypttab_names(devicegraph, "path_to_crypttab")
-
-        expect(device.encryption.method.to_sym).to eq(encryption_method)
-      end
-
-      it "uses the crypttab name for the new encrytion device" do
-        described_class.save_crypttab_names(devicegraph, "path_to_crypttab")
-
-        expect(device.encryption.basename).to eq("luks1")
-      end
-    end
-
-    shared_examples "swap encryption methods" do
-      context "and the crypttab entry matches the random swap encryption method" do
-        let(:encryption_method) { Y2Storage::EncryptionMethod.find(:random_swap) }
-
-        include_examples "swap encryption", :random_swap
-      end
-
-      context "and the crypttab entry matches the protected swap encryption method" do
-        let(:encryption_method) { Y2Storage::EncryptionMethod.find(:protected_swap) }
-
-        include_examples "swap encryption", :protected_swap
-      end
-
-      context "and the crypttab entry matches the secure swap encryption method" do
-        let(:encryption_method) { Y2Storage::EncryptionMethod.find(:secure_swap) }
-
-        include_examples "swap encryption", :secure_swap
-      end
-    end
-
-    context "when the device indicated in a crypttab entry is currently encrypted" do
-      let(:device_name) { "/dev/sda4" }
-
-      before do
-        allow(Y2Storage::EncryptionMethod).to receive(:for_crypttab).and_return(encryption_method)
-      end
-
-      context "and the crypttab entry does not match an encryption method for swap" do
-        let(:encryption_method) { nil }
-
-        it "saves the crypttab name on the encryption device" do
-          encryption_before = device.encryption
-
-          expect(device.encryption.crypttab_name).to be_nil
-
-          described_class.save_crypttab_names(devicegraph, "path_to_crypttab")
-
-          expect(device.encryption.sid).to eq(encryption_before.sid)
-          expect(device.encryption.crypttab_name).to eq("luks1")
-        end
-      end
-
-      include_examples "swap encryption methods"
-    end
-
-    context "when the device indicated in a crypttab entry is not currently encrypted" do
-      let(:device_name) { "/dev/sda1" }
-
-      before do
-        allow(Y2Storage::EncryptionMethod).to receive(:for_crypttab).and_return(encryption_method)
-      end
-
-      context "and the crypttab entry matches none encryption method for swap" do
-        let(:encryption_method) { nil }
-
-        it "does not encrypt the device" do
-          expect(device.encrypted?).to eq(false)
-
-          described_class.save_crypttab_names(devicegraph, "path_to_crypttab")
-
-          expect(device.encrypted?).to eq(false)
-        end
-
-        it "does not fail" do
-          expect { described_class.save_crypttab_names(devicegraph, "path_to_crypttab") }
-            .to_not raise_error
-        end
-      end
-
-      include_examples "swap encryption methods"
-    end
-
-    context "when the device indicated in a crypttab entry is not found" do
-      let(:device_name) { "/dev/sdb1" }
-
-      # Mock the system lookup performed as last resort to find a device
-      before { allow(Y2Storage::BlkDevice).to receive(:find_by_any_name) }
-
-      it "does not fail" do
-        expect { described_class.save_crypttab_names(devicegraph, "path_to_crypttab") }
-          .to_not raise_error
-      end
-    end
-
-    context "when an encrypted device is not indicated in any crypttab entry" do
-      let(:device_name) { "/dev/sda4" }
-
-      it "does not modify the crypttab name of that encryption device" do
-        device = devicegraph.find_by_name("/dev/sda5")
-
-        described_class.save_crypttab_names(devicegraph, "path_to_crypttab")
-
-        expect(device.encryption.crypttab_name).to be_nil
-      end
-    end
-  end
-
   describe ".dm_name_for" do
-    before { fake_scenario(scenario) }
-
-    let(:devicegraph) { Y2Storage::StorageManager.instance.staging }
-
     context "when generating a name for a partition" do
       let(:blk_device) { devicegraph.find_by_name("/dev/sda2") }
